LPS-105162 move 2_0_0 to 2_1_0
